Configuration file initial (manual function)
Combination of initial debug/log/config settings and adjustments in positioning.

.config.json in relative directory will be used temporarily, with no auto-saving.  Can be manually exported in settings tab

JSON is not the ideal format for user-configuration files, but it's an easy one to get started with.  Can add file conversation later (yaml for instance, if added as an install)

Currently only houses the template (auto-load) location, moved from it's hard-coded inline location.