Last part of Infobox merge
I splitted it up into three parts to make sure that I do not accidentally create conflicts with methods added or altered after I copied the version I worked on
This commit (and the two before) contain the new infobox layout. This is only finished for equipment yet, but I'll add the other infoboxes soon. When the infobox in general is finished I will also clean up the setupInfoboxFrame() method [At the moment there is a lot of commented-out code there that I will convert to fit the new layout]